--- devicekit-ios.xcodeproj/tcp/Tests/tcpTests/tcpTests.swift
@@ -1,12 +1,0 @@
-import XCTest
-@testable import TCP
-
-final class tcpTests: XCTestCase {
-    func testExample() throws {
-        // XCTest Documentation
-        // https://developer.apple.com/documentation/xctest
-
-        // Defining Test Cases and Test Methods
-        // https://developer.apple.com/documentation/xctest/defining_test_cases_and_test_methods
-    }
-}
