--- h264-codec 2/Sources/h264-codec/CIImage+Extension.swift
@@ -1,14 +1,0 @@
-import CoreImage
-
-extension CIImage {
-    func flip(orientation: CGImagePropertyOrientation) -> CIImage {
-        switch orientation {
-        case .up, .upMirrored, .down, .downMirrored:
-            return oriented(.up)
-        case .left, .leftMirrored:
-            return oriented(.right)
-        case .right, .rightMirrored:
-            return oriented(.left)
-        }
-    }
-}
