--- h264-codec 2/Tests/h264-codecTests/h264_codecTests.swift
@@ -1,12 +1,0 @@
-import XCTest
-@testable import H264Codec
-
-final class h264_codecTests: XCTestCase {
-    func testExample() throws {
-        // XCTest Documentation
-        // https://developer.apple.com/documentation/xctest
-
-        // Defining Test Cases and Test Methods
-        // https://developer.apple.com/documentation/xctest/defining_test_cases_and_test_methods
-    }
-}
